# openapi/openapi.v1.0.0-rc1.json
{
  "components": {
    "schemas": {
...
      "DealResponse": {
        "properties": {
          "commission": {
            "default": "0",
            "title": "Commission",
            "type": "string"
          },
          "deal_type": {
            "default": "",
            "title": "Deal Type",
            "type": "string"
          },
          "order_ticket": {
            "title": "Order Ticket",
            "type": "integer"
          },
          "price": {
            "title": "Price",
            "type": "string"
          },
          "profit": {
            "title": "Profit",
            "type": "string"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "swap": {
            "default": "0",
            "title": "Swap",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "ticket": {
            "title": "Ticket",
            "type": "integer"
          },
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "ticket",
          "order_ticket",
          "symbol",
          "side",
          "volume",
          "price",
          "profit",
          "time"
        ],
        "title": "DealResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryOrderResponse": {
        "properties": {
          "order_type": {
            "title": "Order Type",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "string"
          },
          "profit": {
            "title": "Profit",
            "type": "string"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "ticket": {
            "title": "Ticket",
            "type": "integer"
          },
          "time_done": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Done"
          },
          "time_setup": {
            "format": "date-time",
            "title": "Time Setup",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "ticket",
          "symbol",
          "side",
          "order_type",
          "volume",
          "price",
          "state",
          "profit",
          "time_setup"
        ],
        "title": "HistoryOrderResponse",
        "type": "object"
      },
      "HistoryResponse": {
        "properties": {
          "deals": {
            "items": {
              "$ref": "#/components/schemas/DealResponse"
            },
            "title": "Deals",
            "type": "array"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/HistoryOrderResponse"
            },
            "title": "Orders",
            "type": "array"
          }
        },
        "title": "HistoryResponse",
        "type": "object"
      },
...
      "PaperHistoryResponse": {
        "properties": {
          "orders": {
            "items": {
              "$ref": "#/components/schemas/PaperOrderResponse"
            },
            "title": "Orders",
            "type": "array"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PaperPositionResponse"
            },
            "title": "Positions",
            "type": "array"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/PaperTradeResponse"
            },
            "title": "Trades",
            "type": "array"
          }
        },
        "required": [
          "orders",
          "trades",
          "positions"
        ],
        "title": "PaperHistoryResponse",
        "type": "object"
      },
...
      "PaperPerformanceResponse": {
        "properties": {
          "performance": {
            "additionalProperties": true,
            "title": "Performance",
            "type": "object"
          },
          "portfolio": {
            "additionalProperties": true,
            "title": "Portfolio",
            "type": "object"
          }
        },
        "required": [
          "performance",
          "portfolio"
        ],
        "title": "PaperPerformanceResponse",
        "type": "object"
      },
      "PaperPositionListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaperPositionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "portfolio": {
            "additionalProperties": true,
            "title": "Portfolio",
            "type": "object"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PaperPositionListResponse",
        "type": "object"
      },
...
      "PendingOrderResponse": {
        "properties": {
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "magic": {
            "default": 0,
            "title": "Magic",
            "type": "integer"
          },
          "order_type": {
            "title": "Order Type",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "string"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "stop_loss": {
            "default": "0",
            "title": "Stop Loss",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "take_profit": {
            "default": "0",
            "title": "Take Profit",
            "type": "string"
          },
          "ticket": {
            "title": "Ticket",
            "type": "integer"
          },
          "volume": {
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "ticket",
          "symbol",
          "side",
          "order_type",
          "volume",
          "price",
          "created_at"
        ],
        "title": "PendingOrderResponse",
        "type": "object"
      },
      "PlacePaperOrderRequest": {
        "properties": {
          "client_order_id": {
            "default": "",
            "maxLength": 64,
            "title": "Client Order Id",
            "type": "string"
          },
          "initial_balance": {
            "default": "10000",
            "description": "Used when creating a new paper portfolio",
            "title": "Initial Balance",
            "type": "string"
          },
          "order_type": {
            "default": "market",
            "description": "market | limit | stop",
            "title": "Order Type",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required for limit/stop orders",
            "title": "Price"
          },
          "reduce_position_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Close/reduce an open paper position",
            "title": "Reduce Position Id"
          },
          "side": {
            "description": "buy | sell",
            "title": "Side",
            "type": "string"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit"
          },
          "volume": {
            "default": "0.10",
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "side"
        ],
        "title": "PlacePaperOrderRequest",
        "type": "object"
      },
      "PlacePaperOrderResponse": {
        "properties": {
          "order": {
            "$ref": "#/components/schemas/PaperOrderResponse"
          },
          "portfolio": {
            "additionalProperties": true,
            "title": "Portfolio",
            "type": "object"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaperPositionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "quote": {
            "additionalProperties": true,
            "title": "Quote",
            "type": "object"
          },
          "trade": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaperTradeResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "order"
        ],
        "title": "PlacePaperOrderResponse",
        "type": "object"
      },
      "PortfolioResponse": {
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AccountSnapshotResponse"
          },
          "history_deals": {
            "items": {
              "$ref": "#/components/schemas/DealResponse"
            },
            "title": "History Deals",
            "type": "array"
          },
          "history_orders": {
            "items": {
              "$ref": "#/components/schemas/HistoryOrderResponse"
            },
            "title": "History Orders",
            "type": "array"
          },
          "pending_order_count": {
            "title": "Pending Order Count",
            "type": "integer"
          },
          "pending_orders": {
            "items": {
              "$ref": "#/components/schemas/PendingOrderResponse"
            },
            "title": "Pending Orders",
            "type": "array"
          },
          "position_count": {
            "title": "Position Count",
            "type": "integer"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionResponse"
            },
            "title": "Positions",
            "type": "array"
          },
          "sync_id": {
            "format": "uuid",
            "title": "Sync Id",
            "type": "string"
          },
          "synced_at": {
            "format": "date-time",
            "title": "Synced At",
            "type": "string"
          }
        },
        "required": [
          "sync_id",
          "account",
          "synced_at",
          "position_count",
          "pending_order_count"
        ],
        "title": "PortfolioResponse",
        "type": "object"
      },
      "PositionResponse": {
        "properties": {
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          },
          "current_price": {
            "title": "Current Price",
            "type": "string"
          },
          "magic": {
            "default": 0,
            "title": "Magic",
            "type": "integer"
          },
          "open_price": {
            "title": "Open Price",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "profit": {
            "title": "Profit",
            "type": "string"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "stop_loss": {
            "title": "Stop Loss",
            "type": "string"
          },
          "swap": {
            "default": "0",
            "title": "Swap",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "take_profit": {
            "title": "Take Profit",
            "type": "string"
          },
          "ticket": {
            "title": "Ticket",
            "type": "integer"
          },
          "volume": {
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "ticket",
          "symbol",
          "side",
          "volume",
          "open_price",
          "current_price",
          "stop_loss",
          "take_profit",
          "profit",
          "opened_at"
        ],
        "title": "PositionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "QuantForg v1.0.0 Release Candidate 1 OpenAPI specification. EXECUTION_ENABLED remains false. No AI endpoints. Generated from FastAPI route registrations.",
    "title": "QuantForg API",
    "version": "1.0.0-rc.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/history": {
      "get": {
        "operationId": "get_history_api_v1_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get History",
        "tags": [
          "portfolio"
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "operationId": "list_orders_api_v1_orders_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PendingOrderResponse"
                  },
                  "title": "Response List Orders Api V1 Orders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Orders",
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/v1/orders/{ticket}": {
      "get": {
        "operationId": "get_order_api_v1_orders__ticket__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket",
            "required": true,
            "schema": {
              "title": "Ticket",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingOrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Order",
        "tags": [
          "portfolio"
        ]
      }
    },
...
    "/api/v1/paper/history": {
      "get": {
        "description": "Paper order / trade / position history.",
        "operationId": "paper_history_api_v1_paper_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Paper History",
        "tags": [
          "paper-trading"
        ]
      }
    },
    "/api/v1/paper/orders": {
      "post": {
        "description": "Place a paper order against live MT5 quotes (simulated fill only).",
        "operationId": "place_paper_order_api_v1_paper_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlacePaperOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacePaperOrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Place Paper Order",
        "tags": [
          "paper-trading"
        ]
      }
    },
    "/api/v1/paper/performance": {
      "get": {
        "description": "Paper portfolio performance snapshot.",
        "operationId": "paper_performance_api_v1_paper_performance_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPerformanceResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Paper Performance",
        "tags": [
          "paper-trading"
        ]
      }
    },
    "/api/v1/paper/positions": {
      "get": {
        "description": "List open paper positions (marked to live MT5 quotes).",
        "operationId": "list_paper_positions_api_v1_paper_positions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPositionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Paper Positions",
        "tags": [
          "paper-trading"
        ]
      }
    },
    "/api/v1/portfolio": {
      "get": {
        "description": "Synchronize and return full portfolio snapshot (read-only).",
        "operationId": "get_portfolio_api_v1_portfolio_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Portfolio",
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/v1/positions": {
      "get": {
        "operationId": "list_positions_api_v1_positions_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PositionResponse"
                  },
                  "title": "Response List Positions Api V1 Positions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Positions",
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/v1/positions/{ticket}": {
      "get": {
        "operationId": "get_position_api_v1_positions__ticket__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket",
            "required": true,
            "schema": {
              "title": "Ticket",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Position",
        "tags": [
          "portfolio"
        ]
      }
    },
...
  }
}